+++ src/app/dashboard/layout.tsx
@@ -1,4 +1,4 @@
-import { DashboardSidebar } from "@/components/dashboard/dashboard-sidebar"
+import { DashboardSidebar } from "@/app/dashboard/_components/dashboard-sidebar"
 import { SidebarProvider } from "@/components/ui/sidebar"
 
 export default function Layout({ children }: { children: React.ReactNode }) {
